# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/go-api/products": {
      "get": {
        "description": "Realiza uma consulta de produtos por paginação",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "List",
        "parameters": [
          {
            "maximum": 500,
            "minimum": 1,
            "type": "integer",
            "default": 100,
            "description": "Quantidade máxima de produtos por página",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Somente produtos públicos",
            "name": "publicOnly",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Página desejada",
            "name": "page",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/products.productsResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/products/{id}": {
      "get": {
        "description": "Realiza uma consulta basica de um produto com base no seu ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "ID",
        "parameters": [
          {
            "type": "integer",
            "description": "ID do produto",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/products.productByIdResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/products/{id}/details": {
      "get": {
        "description": "Realiza uma consulta completa de um produto com base no seu ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Details",
        "parameters": [
          {
            "type": "integer",
            "description": "ID do produto",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/products.productDetailsByIdResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/search/cars": {
      "get": {
        "description": "Realiza uma pesquisa de carros",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Search"
        ],
        "summary": "Cars",
        "parameters": [
          {
            "type": "string",
            "description": "Pesquisa",
            "name": "search",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "limite de resultados (padrão: 100)",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/search.carSearchResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/search/products": {
      "get": {
        "description": "Realiza uma pesquisa de produtos",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Search"
        ],
        "summary": "Products",
        "parameters": [
          {
            "type": "string",
            "description": "Pesquisa",
            "name": "search",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "limite de resultados (padrão: 100)",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/search.productSearchResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/go-api/vehicles": {
      "get": {
        "description": "Realiza uma consulta de veiculos por paginação",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "List",
        "parameters": [
          {
            "maximum": 500,
            "minimum": 1,
            "type": "integer",
            "default": 100,
            "description": "Quantidade máxima de veiculos por página",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Página desejada",
            "name": "page",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/vehicles.vehiclesResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/vehicles/{id}": {
      "get": {
        "description": "Realiza uma consulta basica de um veiculo com base no seu ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "ID",
        "parameters": [
          {
            "type": "integer",
            "description": "ID do veiculo",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/vehicles.vehicleByIdResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    },
    "/go-api/vehicles/{id}/details": {
      "get": {
        "description": "Realiza uma consulta completa de um veiculo com base no seu ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Vehicles"
        ],
        "summary": "Details",
        "parameters": [
          {
            "type": "integer",
            "description": "ID do veiculo",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/vehicles.vehicleDetailsByIdResponse"
            }
          },
          "400": {
            "description": "Requisição inválida",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "schema": {
              "$ref": "#/definitions/httpcustom.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "global.JsonMap": {
      "type": "object",
      "additionalProperties": {}
    },
    "httpcustom.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        }
      }
    },
    "products.application": {
      "type": "object",
      "properties": {
        "engine": {
          "$ref": "#/definitions/products.engine"
        },
        "id": {
          "type": "integer"
        },
        "make": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "transmission": {
          "$ref": "#/definitions/products.transmission"
        },
        "version": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
    "products.engine": {
      "type": "object",
      "properties": {
        "aspiration": {
          "type": "string"
        },
        "cc": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "fuel": {
          "type": "string"
        },
        "valves": {
          "type": "string"
        }
      }
    },
    "products.image": {
      "type": "object",
      "properties": {
        "number": {
          "type": "integer"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "products.product": {
      "type": "object",
      "properties": {
        "appCount": {
          "type": "integer",
          "example": 1
        },
        "code": {
          "type": "string",
          "example": "FAP-2829"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "image": {
          "type": "string",
          "example": "https://www.wega.com.br/wp-content/uploads/2020/09/filtro-de-ar-wega-1.jpg"
        },
        "make": {
          "type": "string",
          "example": "WEGA"
        },
        "name": {
          "type": "string",
          "example": "FILTRO DE AR"
        },
        "public": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "products.productById": {
      "type": "object",
      "properties": {
        "appCount": {
          "type": "integer",
          "example": 1
        },
        "code": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "image": {
          "type": "string",
          "example": "https://www.wega.com.br/wp-content/uploads/2020/09/filtro-de-ar-wega-1.jpg"
        },
        "make": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "products.productByIdResponse": {
      "type": "object",
      "properties": {
        "product": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/products.productById"
          }
        }
      }
    },
    "products.productDetails": {
      "type": "object",
      "properties": {
        "applications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/products.application"
          }
        },
        "code": {
          "type": "string"
        },
        "fiscal_data": {
          "$ref": "#/definitions/global.JsonMap"
        },
        "id": {
          "type": "integer"
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/products.image"
          }
        },
        "logistic_data": {
          "$ref": "#/definitions/global.JsonMap"
        },
        "make": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "similar": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/products.similar"
          }
        },
        "tech_data": {
          "$ref": "#/definitions/global.JsonMap"
        }
      }
    },
    "products.productDetailsByIdResponse": {
      "type": "object",
      "properties": {
        "product": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/products.productDetails"
          }
        }
      }
    },
    "products.productsResponse": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/products.product"
          }
        }
      }
    },
    "products.similar": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "make": {
          "type": "string"
        }
      }
    },
    "products.transmission": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "search.car": {
      "type": "object",
      "properties": {
        "engine": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "make": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "transmission": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
    "search.carSearchResponse": {
      "type": "object",
      "properties": {
        "cars": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/search.car"
          }
        }
      }
    },
    "search.product": {
      "type": "object",
      "properties": {
        "appCount": {
          "type": "integer",
          "example": 1
        },
        "code": {
          "type": "string",
          "example": "FAP-2829"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "image": {
          "type": "string",
          "example": "https://www.wega.com.br/wp-content/uploads/2020/09/filtro-de-ar-wega-1.jpg"
        },
        "make": {
          "type": "string",
          "example": "WEGA"
        },
        "name": {
          "type": "string",
          "example": "FILTRO DE AR"
        }
      }
    },
    "search.productSearchResponse": {
      "type": "object",
      "properties": {
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/search.product"
          }
        }
      }
    },
...
    "vehicles.engine": {
      "type": "object",
      "properties": {
        "aspiration": {
          "type": "string"
        },
        "cc": {
          "type": "integer"
        },
        "code": {
          "type": "string"
        },
        "fuel": {
          "type": "string"
        },
        "valves": {
          "type": "string"
        }
      }
    },
    "vehicles.parts": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "make": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "vehicles.transmission": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "vehicles.vehicle": {
      "type": "object",
      "properties": {
        "engine": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "make": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "partsCount": {
          "type": "integer"
        },
        "transmission": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
    "vehicles.vehicleByIdResponse": {
      "type": "object",
      "properties": {
        "vehicle": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vehicles.vehicle"
          }
        }
      }
    },
    "vehicles.vehicleDetails": {
      "type": "object",
      "properties": {
        "engine": {
          "$ref": "#/definitions/vehicles.engine"
        },
        "id": {
          "type": "integer"
        },
        "info": {
          "$ref": "#/definitions/global.JsonMap"
        },
        "make": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "parts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vehicles.parts"
          }
        },
        "source_url": {
          "type": "string"
        },
        "tech_data": {
          "$ref": "#/definitions/global.JsonMap"
        },
        "transmission": {
          "$ref": "#/definitions/vehicles.transmission"
        },
        "version": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
    "vehicles.vehicleDetailsByIdResponse": {
      "type": "object",
      "properties": {
        "vehicle": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vehicles.vehicleDetails"
          }
        }
      }
    },
    "vehicles.vehiclesResponse": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer"
        },
        "vehicle": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vehicles.vehicle"
          }
        }
      }
    }
  }
}